refactor(consumption): extract TripBaselineRecorder from TripRecording
Pull the #769 / #780 / #894 baseline-learning concern — the largest
cohesive cluster in the notifier — into a focused collaborator: the
per-trip situation classifier, the learned-baseline store, the
classify → record → band → delta pipeline, the fuzzy-calibration
path, and the post-flush server-merge.

The notifier hands one live reading in via recordAndClassify and gets
(situation, band, delta) back; start() delegates the Hive load and
stop() delegates the flush + sync. Six now-unused imports drop off
the notifier.

Completes the "TripRecording notifier split into focused
collaborators" acceptance criterion of #1679.
